Refactor the entire sdk, transfer to async mode.
In the implementation of e2b, it uses threading to perform IO, but in
test I find it is extremely slow. So I decided to implement an all-async
sandbox sdk.

Now almost all functions in sdk are coroutine functions, and the user
need to use asyncio to start their function.

The performance gain is obvious, previously start a process takes more
than 10 seconds, now is almost 1 second or less.